Expanded functionality in placeYourShips()
placeYourShips() now takes into consideration the presence of another ship in generating a list of possible directions for the ship to be placed. List is then printed to the user.

This method could be improved by condensing the two considerations ('is there space on the board?' and 'would there be another ship in the way?') into one iteration.  This would cut the number of iterations through each direction by half in the best case scenario. 

This update was made while listening to the album 'De Cerca' by Josemi Carmona, a great flamenco trio.